requirements.txt
scrapy
sentence-transformers
qdrant-client
customtkinter
python-dotenv
requests
tqdm